% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSigGenes.R
\name{findSigGenes}
\alias{findSigGenes}
\title{findSigGenes}
\usage{
findSigGenes(v, method = "mean")
}
\arguments{
\item{v}{v}

\item{method}{mean or median or abs}
}
\description{
define significant genes along cell.
}
