% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSigGenesGroup.R
\name{findSigGenesGroup}
\alias{findSigGenesGroup}
\title{findSigGenesGroup}
\usage{
findSigGenesGroup(Count = NULL, ClustInfo = NULL, q0 = 0.1)
}
\arguments{
\item{Count}{Count data}

\item{ClustInfo}{}

\item{q0}{qvalue cutoff, default is 0.1}
}
\description{
define significant genes along Group.
}
