% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHyperPvalue.R
\name{getHyperPvalue}
\alias{getHyperPvalue}
\title{Hypergeometric pvalue}
\usage{
getHyperPvalue(genes, genesets)
}
\arguments{
\item{genes}{genes}
}
\description{
calculate hypergeometric pvalue for given genes
}
