% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GK.R
\name{GK}
\alias{GK}
\title{Gaussian Kernal transform (GSVA)}
\usage{
GK(v)
}
\arguments{
\item{v}{CountData [dgCMatrix]}
}
\value{
transformed CountData
}
\description{
Transform count data with gaussian kernal using pre-calculated cdf
This takes ~ 3 min with 20000 * 90 size datas, (Really Really Slow)
}
\seealso{
https://bmcbioinformatics.biomedcentral.com/articles/10.1186/1471-2105-14-7
}
