% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSigGenes.R
\name{findSigGenes}
\alias{findSigGenes}
\title{findSigGenes}
\usage{
findSigGenes(v, method = "median")
}
\arguments{
\item{v}{count data}

\item{method}{mean or pos}
}
\description{
define significant genes along cell.
}
