% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findSigGenesGroup.R
\name{findSigGenesGroup}
\alias{findSigGenesGroup}
\title{findSigGenesGroup}
\usage{
findSigGenesGroup(Count = NULL, ClustInfo = NULL, q0 = 0.1, TopCutoff = 5)
}
\arguments{
\item{Count}{Count data}

\item{ClustInfo}{Group information}

\item{q0}{qvalue cutoff, default is 0.1}

\item{TopCutoff}{cutoff for Top in findMarkers, default is 5.}
}
\description{
define significant genes along Group.
}
