% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GK.R
\name{GK}
\alias{GK}
\title{gaussian kernal transform (GSVA)}
\usage{
GK(v)
}
\arguments{
\item{v}{count data}
}
\description{
transform count data with gaussian kernal using pre calculated cdf
this takes ~ 3 min with 20000 * 90 size datas
}
\seealso{
https://bmcbioinformatics.biomedcentral.com/articles/10.1186/1471-2105-14-7
}
